pruebas para mandar los mensajes de notificaciones al main contact
