+++ whatsapp.js
@@ -5,10 +5,13 @@ const QRCode = require('qrcode-terminal');
 const googleApi = require('./googleApi'); // Importa el módulo entero
 const { credentials } = require('./config');
 const fs = require('fs');
+const { getProfileName } = require('./utils'); // Importa getProfileName desde utils.js
+const { listEvents } = require('./listEvents'); // Importa listEvents desde listEvents.js
 
 const CONTACTED_USERS_FILE = './contactedUsers.json';
 const MAIN_CONTACT = '5491126320824@s.whatsapp.net';
 let contactedUsers = [];
+let invalidMeetingsDetails = ''; // Variable para almacenar detalles de reuniones no válidas
 
 // Load contacted users from file
 function loadContactedUsers() {
@@ -32,8 +35,6 @@ function addUserToContacted(userId, eventId) {
     contactedUsers.push({ userId, eventId });
     saveContactedUsers();
 }
-
-let upcomingEvents = []; // Inicializamos como un array vacío
 
 async function sendAuthUrl(authUrl) {
     const { state, saveCreds } = await useMultiFileAuthState('auth_info_baileys');
@@ -116,24 +117,25 @@ async function connectToWhatsApp(oAuth2Client) {
         if (messageType === 'conversation' || messageType === 'extendedTextMessage') {
             if (messageContent.toLowerCase() === 'start') {
                 await sock.sendMessage(MAIN_CONTACT, { text: 'Comando "start" recibido. Chequeando eventos y enviando mensajes...' });
-                await notifyNextEvents(sock, userId, oAuth2Client);
+                await notifyNextEvents(sock, userId, oAuth2Client, getProfileName);
             }
         }
     });
 
-    async function notifyNextEvents(sock, userId, authClient) {
-        const profileName = await googleApi.getProfileName(authClient);
-        const upcomingEvents = await googleApi.listEvents(authClient, profileName, sock) || []; // Pasar profileName y sock
+    async function notifyNextEvents(sock, userId, authClient, getProfileName) {
+        const profileName = await getProfileName(authClient);
+        const upcomingEvents = await listEvents(authClient, profileName, sock) || []; // Pasar profileName y sock
+    
+        let sentMessagesCount = 0; // Contador de mensajes enviados
+        let sentMessagesDetails = ''; // Detalles de los mensajes enviados
+        let notSentMessagesDetails = ''; // Detalles de los mensajes no enviados
+        let invalidMeetingsDetails = ''; // Detalles de las reuniones no válidas
     
         if (upcomingEvents.length > 0) {
-            let sentMessagesCount = 0; // Contador de mensajes enviados
-            let sentMessagesDetails = ''; // Detalles de los mensajes enviados
-            let notSentMessagesDetails = ''; // Detalles de los mensajes no enviados
-    
             for (const event of upcomingEvents) {
                 if (!isUserContacted(event.userId, event.eventId)) {
                     const dynamicMessagePart = event.dynamicMessagePart; // Obtener mensaje dinámico ya procesado
-                    
+    
                     // Mensajes dependiendo del país
                     let messages = [];
                     if (event.country === 'Canada/EEUU') {
@@ -161,16 +163,31 @@ async function connectToWhatsApp(oAuth2Client) {
                     notSentMessagesDetails += `No se envió mensaje a ${event.userId.split('@')[0]} porque ya se había enviado anteriormente\n`;
                 }
             }
-            let consolidatedMessage = `Mensajes enviados a ${sentMessagesCount} próximo(s) evento(s).\n\n${sentMessagesDetails}\n${notSentMessagesDetails}`;
-            await sock.sendMessage(MAIN_CONTACT, { text: consolidatedMessage });
         } else {
             console.log('No se encontraron eventos.');
             await sock.sendMessage(MAIN_CONTACT, { text: 'No se encontraron eventos.' });
         }
+    
+        // Leer números no registrados desde el archivo
+        let unregisteredNumbers = googleApi.loadUnregisteredNumbers();
+    
+        // Consolidar y enviar el mensaje final
+        let consolidatedMessage = `Mensajes enviados a ${sentMessagesCount} próximo(s) evento(s).\n\n${sentMessagesDetails}\n${notSentMessagesDetails}`;
+    
+        if (unregisteredNumbers.length > 0) {
+            consolidatedMessage += `\n\nNúmeros no registrados en WhatsApp:\n${unregisteredNumbers.join('\n')}`;
+        }
+    
+        if (invalidMeetingsDetails.length > 0) {
+            consolidatedMessage += `\n\nReuniones no válidas:\n${invalidMeetingsDetails}`;
+        }
+    
+        await sock.sendMessage(MAIN_CONTACT, { text: consolidatedMessage });
+        // Resetear el archivo de números no registrados después de enviar el mensaje
+        googleApi.saveUnregisteredNumbers([]);
     }
 
     // Ya no se necesita notifyUnregisteredNumbers aquí
-
 }
 
 module.exports = { sendAuthUrl, connectToWhatsApp };
